Add bcrypt dependency and refactor(modularisation) user validation middleware
