+++ backend/package.json
@@ -10,6 +10,7 @@
   "author": "the_demon_sid",
   "license": "ISC",
   "dependencies": {
+    "bcrypt": "^5.1.1",
     "dotenv": "^16.4.5",
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.2",
