+++ backend/src/models/users.model.js
@@ -16,8 +16,6 @@ const userSchema = new mongoose.Schema({
   password: {
     type: String,
     required: true,
-    minlength: 8,
-    maxlength: 16,
     unique: false,
   },
 });
